Added unlimited multiple font rendering support. Added wide string rendering suppport. Changed the buffer interface to use size_t instead of int. Moved uniform, sampler and texture data from VisualPass (renamed as VisualEffectPass) into VisualEffectInstance.

+++ include/FalconEngine/Graphics/Renderer/Resources/IndexBuffer.h
@@ -22,7 +22,7 @@ public:
     /************************************************************************/
     /* Constructors and Destructor                                          */
     /************************************************************************/
-    IndexBuffer(int indexNum, IndexType indexType, BufferUsage usage);
+    IndexBuffer(size_t indexNum, IndexType indexType, BufferUsage usage);
     virtual ~IndexBuffer();
 
 public:
